+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 All notable changes to the "MeerkatIO" extension will be documented in this file.
 
+## [1.5.6]
+
+- Add survey prompt for extension feedback and improvement
+
 ## [1.5.5]
 
--Improve terminal watcher to account for the currently active terminal on extension activation
+- Improve terminal watcher to account for the currently active terminal on extension activation
 
 ## [1.5.4]